+++ scripts/lib/setup_venv.py
@@ -28,6 +28,7 @@ VENV_DEPENDENCIES = [
     "jq",  # No longer used in production (clean me up later)
     "libsasl2-dev",  # For building python-ldap from source
     "libvips",  # For thumbnailing
+    "libvips-tools",
 ]
 
 COMMON_YUM_VENV_DEPENDENCIES = [
@@ -45,7 +46,8 @@ COMMON_YUM_VENV_DEPENDENCIES = [
     "postgresql-libs",  # libpq-dev on apt
     "openssl-devel",
     "jq",
-    "libvips",  # For thumbnailing
+    "vips",  # For thumbnailing
+    "vips-tools",
 ]
 
 REDHAT_VENV_DEPENDENCIES = [