+++ web/src/info_overlay.ts
@@ -204,12 +204,12 @@ Last task`,
     <div class="widget-content">
         <div class="todo-widget">
             <h4>Today's tasks</h4>
-            <ul class="todo-widget new-style">
+            <ul class="todo-widget">
                 <li>
                     <label class="checkbox">
                         <div>
                             <input type="checkbox" class="task" checked="checked">
-                            <span></span>
+                            <span class="rendered-checkbox"></span>
                         </div>
                         <div>
                             <strike><strong>Task 1:</strong> This is the first task.</strike>
@@ -220,7 +220,7 @@ Last task`,
                     <label class="checkbox">
                         <div>
                             <input type="checkbox" class="task">
-                            <span></span>
+                            <span class="rendered-checkbox"></span>
                         </div>
                         <div>
                             <strong>Task 2:</strong> This is the second task.
@@ -231,7 +231,7 @@ Last task`,
                     <label class="checkbox">
                         <div>
                             <input type="checkbox" class="task">
-                            <span></span>
+                            <span class="rendered-checkbox"></span>
                         </div>
                         <div>
                             <strong>Last task</strong>