+++ zerver/tests/test_user_groups.py
@@ -8,7 +8,15 @@ from django.db import transaction
 from django.utils.timezone import now as timezone_now
 
 from zerver.actions.create_realm import do_create_realm
-from zerver.actions.realm_settings import do_set_realm_property
+from zerver.actions.realm_settings import (
+    do_change_realm_permission_group_setting,
+    do_set_realm_property,
+)
+from zerver.actions.streams import (
+    do_change_stream_group_based_setting,
+    do_deactivate_stream,
+    do_unarchive_stream,
+)
 from zerver.actions.user_groups import (
     add_subgroups_to_user_group,
     bulk_add_members_to_user_groups,
@@ -16,6 +24,7 @@ from zerver.actions.user_groups import (
     check_add_user_group,
     create_user_group_in_database,
     do_change_user_group_permission_setting,
+    do_deactivate_user_group,
     promote_new_full_members,
 )
 from zerver.actions.users import do_deactivate_user
@@ -43,6 +52,7 @@ from zerver.models import (
     GroupGroupMembership,
     NamedUserGroup,
     Realm,
+    Stream,
     UserGroup,
     UserGroupMembership,
     UserProfile,
@@ -74,7 +84,7 @@ class UserGroupTestCase(ZulipTestCase):
         assert user_group is not None
         empty_user_group = check_add_user_group(realm, "newgroup", [], acting_user=None)
 
-        user_groups = user_groups_in_realm_serialized(realm)
+        user_groups = user_groups_in_realm_serialized(realm, allow_deactivated=False)
         self.assert_length(user_groups, 10)
         self.assertEqual(user_groups[0]["id"], user_group.id)
         self.assertEqual(user_groups[0]["name"], SystemGroups.NOBODY)
@@ -83,6 +93,7 @@ class UserGroupTestCase(ZulipTestCase):
         self.assertEqual(user_groups[0]["direct_subgroup_ids"], [])
         self.assertEqual(user_groups[0]["can_manage_group"], user_group.id)
         self.assertEqual(user_groups[0]["can_mention_group"], user_group.id)
+        self.assertFalse(user_groups[0]["deactivated"])
 
         owners_system_group = NamedUserGroup.objects.get(name=SystemGroups.OWNERS, realm=realm)
         membership = UserGroupMembership.objects.filter(user_group=owners_system_group).values_list(
@@ -95,6 +106,7 @@ class UserGroupTestCase(ZulipTestCase):
         self.assertEqual(user_groups[1]["direct_subgroup_ids"], [])
         self.assertEqual(user_groups[1]["can_manage_group"], user_group.id)
         self.assertEqual(user_groups[1]["can_mention_group"], user_group.id)
+        self.assertFalse(user_groups[0]["deactivated"])
 
         admins_system_group = NamedUserGroup.objects.get(
             name=SystemGroups.ADMINISTRATORS, realm=realm
@@ -112,6 +124,7 @@ class UserGroupTestCase(ZulipTestCase):
         self.assertEqual(user_groups[9]["members"], [])
         self.assertEqual(user_groups[9]["can_manage_group"], user_group.id)
         self.assertEqual(user_groups[9]["can_mention_group"], everyone_group.id)
+        self.assertFalse(user_groups[0]["deactivated"])
 
         othello = self.example_user("othello")
         hamletcharacters_group = NamedUserGroup.objects.get(name="hamletcharacters", realm=realm)
@@ -128,7 +141,7 @@ class UserGroupTestCase(ZulipTestCase):
             },
             acting_user=None,
         )
-        user_groups = user_groups_in_realm_serialized(realm)
+        user_groups = user_groups_in_realm_serialized(realm, allow_deactivated=False)
         self.assertEqual(user_groups[10]["id"], new_user_group.id)
         self.assertEqual(user_groups[10]["name"], "newgroup2")
         self.assertEqual(user_groups[10]["description"], "")
@@ -146,6 +159,35 @@ class UserGroupTestCase(ZulipTestCase):
         self.assertCountEqual(
             user_groups[10]["can_mention_group"].direct_subgroups,
             [admins_system_group.id, hamletcharacters_group.id],
+        )
+        self.assertFalse(user_groups[0]["deactivated"])
+
+        another_new_group = check_add_user_group(
+            realm, "newgroup3", [self.example_user("hamlet")], acting_user=None
+        )
+        add_subgroups_to_user_group(
+            new_user_group, [another_new_group, owners_system_group], acting_user=None
+        )
+        do_deactivate_user_group(another_new_group, acting_user=None)
+        user_groups = user_groups_in_realm_serialized(realm, allow_deactivated=True)
+        self.assert_length(user_groups, 12)
+        self.assertEqual(user_groups[10]["id"], new_user_group.id)
+        self.assertEqual(user_groups[10]["name"], "newgroup2")
+        self.assertFalse(user_groups[10]["deactivated"])
+        self.assertCountEqual(
+            user_groups[10]["direct_subgroup_ids"], [another_new_group.id, owners_system_group.id]
+        )
+        self.assertEqual(user_groups[11]["id"], another_new_group.id)
+        self.assertEqual(user_groups[11]["name"], "newgroup3")
+        self.assertTrue(user_groups[11]["deactivated"])
+
+        user_groups = user_groups_in_realm_serialized(realm, allow_deactivated=False)
+        self.assert_length(user_groups, 11)
+        self.assertEqual(user_groups[10]["id"], new_user_group.id)
+        self.assertEqual(user_groups[10]["name"], "newgroup2")
+        self.assertFalse(user_groups[10]["deactivated"])
+        self.assertCountEqual(
+            user_groups[10]["direct_subgroup_ids"], [another_new_group.id, owners_system_group.id]
         )
 
     def test_get_direct_user_groups(self) -> None:
@@ -606,6 +648,35 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_post("/json/user_groups/create", info=params)
         self.assert_json_error(result, "Invalid user group ID: 1111")
 
+        # Test can_mention_group cannot be set to a deactivated group.
+        do_deactivate_user_group(leadership_group, acting_user=None)
+        params = {
+            "name": "social",
+            "members": orjson.dumps([hamlet.id]).decode(),
+            "description": "Social team",
+            "can_mention_group": orjson.dumps(leadership_group.id).decode(),
+        }
+        result = self.client_post("/json/user_groups/create", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+
+        params = {
+            "name": "social",
+            "members": orjson.dumps([hamlet.id]).decode(),
+            "description": "Social team",
+            "can_mention_group": orjson.dumps(
+                {
+                    "direct_members": [othello.id],
+                    "direct_subgroups": [leadership_group.id],
+                }
+            ).decode(),
+        }
+        result = self.client_post("/json/user_groups/create", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+
+        # Reactivate group to use it in further tests.
+        leadership_group.deactivated = False
+        leadership_group.save()
+
         with self.settings(ALLOW_GROUP_VALUED_SETTINGS=False):
             params = {
                 "name": "frontend",
@@ -746,6 +817,17 @@ class UserGroupAPITestCase(UserGroupTestCase):
         params = {"name": "channel:1"}
         result = self.client_patch(f"/json/user_groups/{user_group.id}", info=params)
         self.assert_json_error(result, "User group name cannot start with 'channel:'.")
+
+        do_deactivate_user_group(user_group, acting_user=None)
+        params = {"description": "Troubleshooting and support team"}
+        result = self.client_patch(f"/json/user_groups/{user_group.id}", info=params)
+        self.assert_json_error(result, "You can only change name of deactivated user groups")
+
+        params = {"name": "Support team"}
+        result = self.client_patch(f"/json/user_groups/{user_group.id}", info=params)
+        self.assert_json_success(result)
+        user_group = NamedUserGroup.objects.get(id=user_group.id)
+        self.assertEqual(user_group.name, "Support team")
 
     def do_test_update_user_group_permission_settings(self, setting_name: str) -> None:
         hamlet = self.example_user("hamlet")
@@ -894,6 +976,31 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_patch(f"/json/user_groups/{support_group.id}", info=params)
         self.assert_json_error(result, "Invalid user group ID: 1111")
 
+        leadership_group = NamedUserGroup.objects.get(realm=hamlet.realm, name="leadership")
+        do_deactivate_user_group(leadership_group, acting_user=None)
+
+        params[setting_name] = orjson.dumps({"new": leadership_group.id}).decode()
+        result = self.client_patch(f"/json/user_groups/{support_group.id}", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+
+        params[setting_name] = orjson.dumps(
+            {
+                "new": {
+                    "direct_members": [prospero.id],
+                    "direct_subgroups": [leadership_group.id],
+                }
+            }
+        ).decode()
+        result = self.client_patch(f"/json/user_groups/{support_group.id}", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+
+        params[setting_name] = orjson.dumps({"new": moderators_group.id}).decode()
+        result = self.client_patch(f"/json/user_groups/{leadership_group.id}", info=params)
+        self.assert_json_error(result, "You can only change name of deactivated user groups")
+
+        leadership_group.deactivated = False
+        leadership_group.save()
+
         # Test case when ALLOW_GROUP_VALUED_SETTINGS is False.
         with self.settings(ALLOW_GROUP_VALUED_SETTINGS=False):
             params[setting_name] = orjson.dumps(
@@ -938,6 +1045,7 @@ class UserGroupAPITestCase(UserGroupTestCase):
         hamlet = self.example_user("hamlet")
         check_add_user_group(hamlet.realm, "support", [hamlet], acting_user=None)
         check_add_user_group(hamlet.realm, "marketing", [hamlet], acting_user=None)
+        check_add_user_group(hamlet.realm, "leadership", [hamlet], acting_user=None)
 
         for setting_name in NamedUserGroup.GROUP_PERMISSION_SETTINGS:
             self.do_test_update_user_group_permission_settings(setting_name)
@@ -1149,6 +1257,238 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_delete(f"/json/user_groups/{lear_test_group.id}")
         self.assert_json_error(result, "Invalid user group")
 
+    def test_user_group_deactivation(self) -> None:
+        support_group = self.create_user_group_for_test("support")
+        leadership_group = self.create_user_group_for_test("leadership")
+        add_subgroups_to_user_group(support_group, [leadership_group], acting_user=None)
+        realm = get_realm("zulip")
+
+        do_set_realm_property(
+            realm, "user_group_edit_policy", CommonPolicyEnum.ADMINS_ONLY, acting_user=None
+        )
+        self.login("othello")
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_error(result, "Insufficient permission")
+
+        do_set_realm_property(
+            realm, "user_group_edit_policy", CommonPolicyEnum.MEMBERS_ONLY, acting_user=None
+        )
+
+        self.login("hamlet")
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_error(result, "Insufficient permission")
+
+        self.login("othello")
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_success(result)
+        support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+        self.assertTrue(support_group.deactivated)
+
+        support_group.deactivated = False
+        support_group.save()
+
+        # Check admins can deactivate groups even if they are not members
+        # of the group.
+        self.login("iago")
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_success(result)
+        support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+        self.assertTrue(support_group.deactivated)
+
+        support_group.deactivated = False
+        support_group.save()
+
+        # Check moderators can deactivate groups if they are allowed by
+        # user_group_edit_policy even when they are not members of the group.
+        do_set_realm_property(
+            realm, "user_group_edit_policy", CommonPolicyEnum.ADMINS_ONLY, acting_user=None
+        )
+        self.login("shiva")
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_error(result, "Insufficient permission")
+
+        do_set_realm_property(
+            realm, "user_group_edit_policy", CommonPolicyEnum.MODERATORS_ONLY, acting_user=None
+        )
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_success(result)
+        support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+        self.assertTrue(support_group.deactivated)
+
+        support_group.deactivated = False
+        support_group.save()
+
+        # Check that group that is subgroup of another group cannot be deactivated.
+        result = self.client_post(f"/json/user_groups/{leadership_group.id}/deactivate")
+        self.assert_json_error(
+            result, "You cannot deactivate a user group that is subgroup of any user group."
+        )
+
+        # If the supergroup is itself deactivated, then subgroup can be deactivated.
+        result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+        self.assert_json_success(result)
+        result = self.client_post(f"/json/user_groups/{leadership_group.id}/deactivate")
+        self.assert_json_success(result)
+        leadership_group = NamedUserGroup.objects.get(name="leadership", realm=realm)
+        self.assertTrue(leadership_group.deactivated)
+
+        # Check that system groups cannot be deactivated at all.
+        self.login("desdemona")
+        members_system_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MEMBERS, realm=realm, is_system_group=True
+        )
+        result = self.client_post(f"/json/user_groups/{members_system_group.id}/deactivate")
+        self.assert_json_error(result, "Insufficient permission")
+
+    def test_user_group_deactivation_with_group_used_for_settings(self) -> None:
+        support_group = self.create_user_group_for_test("support")
+        realm = get_realm("zulip")
+        moderators_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MODERATORS, realm=realm, is_system_group=True
+        )
+        hamlet = self.example_user("hamlet")
+        self.login("desdemona")
+
+        for setting_name in Realm.REALM_PERMISSION_GROUP_SETTINGS:
+            anonymous_setting_group = self.create_or_update_anonymous_group_for_setting(
+                [hamlet], [moderators_group, support_group]
+            )
+            do_change_realm_permission_group_setting(
+                realm, setting_name, anonymous_setting_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            do_change_realm_permission_group_setting(
+                realm, setting_name, support_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            # Reset the realm setting to one of the system group so this setting
+            # does not interfere when testing for another setting.
+            do_change_realm_permission_group_setting(
+                realm, setting_name, moderators_group, acting_user=None
+            )
+
+        stream = ensure_stream(realm, "support", acting_user=None)
+        for setting_name in Stream.stream_permission_group_settings:
+            do_change_stream_group_based_setting(
+                stream, setting_name, support_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            # Test the group can be deactivated, if the stream which uses
+            # this group for a setting is deactivated.
+            do_deactivate_stream(stream, acting_user=None)
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_success(result)
+            support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+            self.assertTrue(support_group.deactivated)
+
+            support_group.deactivated = False
+            support_group.save()
+
+            do_unarchive_stream(stream, "support", acting_user=None)
+
+            anonymous_setting_group = self.create_or_update_anonymous_group_for_setting(
+                [hamlet], [moderators_group, support_group]
+            )
+            do_change_stream_group_based_setting(
+                stream, setting_name, anonymous_setting_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            # Test the group can be deactivated, if the stream which uses
+            # this group for a setting is deactivated.
+            do_deactivate_stream(stream, acting_user=None)
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_success(result)
+            support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+            self.assertTrue(support_group.deactivated)
+
+            # Reactivate the group again for further testing.
+            support_group.deactivated = False
+            support_group.save()
+
+            # Reset the stream setting to one of the system group so this setting
+            # does not interfere when testing for another setting.
+            do_change_stream_group_based_setting(
+                stream, setting_name, moderators_group, acting_user=None
+            )
+
+        leadership_group = self.create_user_group_for_test("leadership")
+        for setting_name in NamedUserGroup.GROUP_PERMISSION_SETTINGS:
+            do_change_user_group_permission_setting(
+                leadership_group, setting_name, support_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            # Test the group can be deactivated, if the user group which uses
+            # this group for a setting is deactivated.
+            do_deactivate_user_group(leadership_group, acting_user=None)
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_success(result)
+            support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+            self.assertTrue(support_group.deactivated)
+
+            support_group.deactivated = False
+            support_group.save()
+
+            leadership_group.deactivated = False
+            leadership_group.save()
+
+            anonymous_setting_group = self.create_or_update_anonymous_group_for_setting(
+                [hamlet], [moderators_group, support_group]
+            )
+            do_change_user_group_permission_setting(
+                leadership_group, setting_name, anonymous_setting_group, acting_user=None
+            )
+
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_error(
+                result, "You cannot deactivate a user group which is used for setting."
+            )
+
+            # Test the group can be deactivated, if the user group which uses
+            # this group for a setting is deactivated.
+            do_deactivate_user_group(leadership_group, acting_user=None)
+            result = self.client_post(f"/json/user_groups/{support_group.id}/deactivate")
+            self.assert_json_success(result)
+            support_group = NamedUserGroup.objects.get(name="support", realm=realm)
+            self.assertTrue(support_group.deactivated)
+
+            # Reactivate the group again for further testing.
+            support_group.deactivated = False
+            support_group.save()
+
+            leadership_group.deactivated = False
+            leadership_group.save()
+
+            # Reset the group setting to one of the system group so this setting
+            # does not interfere when testing for another setting.
+            do_change_user_group_permission_setting(
+                leadership_group, setting_name, moderators_group, acting_user=None
+            )
+
     def test_query_counts(self) -> None:
         hamlet = self.example_user("hamlet")
         cordelia = self.example_user("cordelia")
@@ -1294,6 +1634,19 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_post(f"/json/user_groups/{user_group.id}/members", info={})
         msg = 'Nothing to do. Specify at least one of "add" or "delete".'
         self.assert_json_error(result, msg)
+        self.assert_user_membership(user_group, [hamlet])
+
+        # Test adding or removing members from a deactivated group.
+        do_deactivate_user_group(user_group, acting_user=None)
+
+        params = {"delete": orjson.dumps([hamlet.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{user_group.id}/members", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+        self.assert_user_membership(user_group, [hamlet])
+
+        params = {"add": orjson.dumps([iago.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{user_group.id}/members", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
         self.assert_user_membership(user_group, [hamlet])
 
     def test_mentions(self) -> None:
@@ -2182,6 +2535,31 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info={})
         self.assert_json_error(result, 'Nothing to do. Specify at least one of "add" or "delete".')
         self.assert_subgroup_membership(support_group, [leadership_group])
+
+        # Do not have support group as subgroup of any group to follow
+        # the condition a group used as a subgroup cannot be deactivated.
+        params = {"delete": orjson.dumps([support_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{test_group.id}/subgroups", info=params)
+        self.assert_json_success(result)
+        self.assert_subgroup_membership(test_group, [])
+
+        # Test adding or removing subgroups from a deactivated group.
+        do_deactivate_user_group(support_group, acting_user=None)
+
+        params = {"delete": orjson.dumps([leadership_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+        self.assert_subgroup_membership(support_group, [leadership_group])
+
+        params = {"add": orjson.dumps([test_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
+        self.assert_subgroup_membership(support_group, [leadership_group])
+
+        # Test that a deactivated group cannot be used as a subgroup.
+        params = {"add": orjson.dumps([support_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{test_group.id}/subgroups", info=params)
+        self.assert_json_error(result, "User group is deactivated.")
 
     def test_get_is_user_group_member_status(self) -> None:
         self.login("iago")