+++ zproject/urls.py
@@ -188,12 +188,13 @@ from zerver.views.upload import (
 )
 from zerver.views.user_groups import (
     add_user_group,
+    deactivate_user_group,
     delete_user_group,
     edit_user_group,
     get_is_user_group_member,
     get_subgroups_of_user_group,
-    get_user_group,
     get_user_group_members,
+    get_user_groups,
     update_subgroups_of_user_group,
     update_user_group_backend,
 )
@@ -397,7 +398,7 @@ v1_api_and_json_patterns = [
     rest_path("users/me/status", POST=update_user_status_backend),
     rest_path("users/<int:user_id>/status", GET=get_status_backend),
     # user_groups -> zerver.views.user_groups
-    rest_path("user_groups", GET=get_user_group),
+    rest_path("user_groups", GET=get_user_groups),
     rest_path("user_groups/create", POST=add_user_group),
     rest_path("user_groups/<int:user_group_id>", PATCH=edit_user_group, DELETE=delete_user_group),
     rest_path(
@@ -413,6 +414,7 @@ v1_api_and_json_patterns = [
     rest_path(
         "user_groups/<int:user_group_id>/members/<int:user_id>", GET=get_is_user_group_member
     ),
+    rest_path("user_groups/<int:user_group_id>/deactivate", POST=deactivate_user_group),
     # users/me -> zerver.views.user_settings
     rest_path("users/me/avatar", POST=set_avatar_backend, DELETE=delete_avatar_backend),
     # users/me/onboarding_steps -> zerver.views.onboarding_steps